refactor(tsjs): split extractor by concern, cover handlers, decompose walk
Split tsjs.go (1212 lines) into a walk-core plus frameworks.go (the
Stimulus controller idiom) and imports.go (dynamic import, re-exports, and
module-constant tracking), with tests relocated to matching files. The
walker's qualified-name construction collapses into a single qualify()
helper.

- Lift the 66-75% handler cluster (Stimulus, function, enum, type alias,
  dynamic import, re-export, composes) to 86-100%; package 89.3 -> 95.1%.
- Add JavaScript golden fixtures (exports.js, imports.js) for the JS-only
  default-export and dynamic-import paths a TS-only set leaves unobserved.
- Decompose walk, collectModuleConstants, and collectExportedNames into
  named helpers, retiring all three complexity-ledger entries.
- The 22 existing goldens stay byte-identical (no -update); behavior
  unchanged.
