+++ internal/extract/testdata/javascript/exports.golden.json
@@ -1,0 +1,68 @@
+{
+  "symbols": [
+    {
+      "name": "Exports",
+      "qualified": "Exports",
+      "kind": "class",
+      "visibility": "public",
+      "line_start": 12,
+      "line_end": 16
+    },
+    {
+      "name": "run",
+      "qualified": "Exports.run",
+      "kind": "method",
+      "visibility": "public",
+      "parent": "Exports",
+      "line_start": 13,
+      "line_end": 15
+    },
+    {
+      "name": "VERSION",
+      "qualified": "VERSION",
+      "kind": "constant",
+      "visibility": "public",
+      "line_start": 8,
+      "line_end": 8
+    },
+    {
+      "name": "internalCache",
+      "qualified": "internalCache",
+      "kind": "constant",
+      "visibility": "private",
+      "line_start": 10,
+      "line_end": 10
+    },
+    {
+      "name": "setup",
+      "qualified": "setup",
+      "kind": "function",
+      "visibility": "public",
+      "line_start": 4,
+      "line_end": 6
+    }
+  ],
+  "edges": [
+    {
+      "source": "Exports",
+      "target": "Base",
+      "kind": "inherits",
+      "line": 12,
+      "confidence": 1
+    },
+    {
+      "source": "Exports.run",
+      "target": "setup",
+      "kind": "calls",
+      "line": 14,
+      "confidence": 1
+    },
+    {
+      "source": "setup",
+      "target": "configure",
+      "kind": "calls",
+      "line": 5,
+      "confidence": 1
+    }
+  ]
+}
