+++ internal/extract/testdata/javascript/imports.golden.json
@@ -1,0 +1,58 @@
+{
+  "symbols": [
+    {
+      "name": "Button",
+      "qualified": "Button",
+      "kind": "constant",
+      "visibility": "public",
+      "line_start": 4,
+      "line_end": 4
+    },
+    {
+      "name": "lazyPanel",
+      "qualified": "lazyPanel",
+      "kind": "function",
+      "visibility": "public",
+      "line_start": 12,
+      "line_end": 12
+    },
+    {
+      "name": "load",
+      "qualified": "load",
+      "kind": "function",
+      "visibility": "public",
+      "line_start": 7,
+      "line_end": 10
+    }
+  ],
+  "edges": [
+    {
+      "source": "",
+      "target": "./button",
+      "kind": "imports",
+      "line": 4,
+      "confidence": 1
+    },
+    {
+      "source": "",
+      "target": "./heavy",
+      "kind": "imports",
+      "line": 8,
+      "confidence": 0.8
+    },
+    {
+      "source": "",
+      "target": "./panel",
+      "kind": "imports",
+      "line": 12,
+      "confidence": 0.8
+    },
+    {
+      "source": "",
+      "target": "./widgets",
+      "kind": "imports",
+      "line": 5,
+      "confidence": 1
+    }
+  ]
+}
